Add Ribs responsive grid system
One of the big points of this redesign is to be mobile-friendly, and I don't
want to write that code myself.  However, many responsive frameworks include
all sorts of opinionated style decisions, too - and I want to do that myself.

Ribs is a maintained fork of Skeleton, which seems reasonably popular.
Notably, it adds support for variables for all its configuration, which should
make things easier.

However, that configuration is in the project itself, which makes using
submodules not possible... this is v1.0.12.